refactor: update imports for AutoPositionSelector and AutoShotInfoSection components
